more tests tidy
